If polygonHull has only 2 points, average them to get centroid
I don't know whether this is a thing that can mess up the centroid calculation.
But I found the same check in the ModeRotate, and it was maybe added for good
reason, so we might as well add this check here too.

We do prevent users from generating 2-point areas, but you never know what
kind of degenerate geometries we may fetch from OSM and other data sources.