Working on misdetection of mousewheel zoom vs pan (re: #680)
This commit contains a bunch of cleanups in the event detection code
and related comments and documentation.
It probably doesn't fix the issue, but shouldn't make it worse either.

In summary:
The old v1 code default to "zoom", but would "pan" in some situations.
The buggy v2-beta code was default to "pan", but "zoom" in some situations.

The detection for whether to choose zoom or pan depended a lot on
whether the delta values where round number or fractional number.

On trackpads this works, however actual mousewheel zooms can have
round or fractional values - this can't be the deciding factor.

Going to test with Windows/Linux and other browsers some more to see
whether there is a more reliable method for detecting mouse 'wheel'
events (which should always "zoom" and never "pan")

Failing that, I'll just make it a configurable option.